model/requirements.txt
torch
torchvision
scikit-learn
seaborn
matplotlib
tqdm
Pillow
psutil
py-cpuinfo
